Rama

+++ app/Http/Controllers/Admin/AdminSlikController.php
@@ -46,13 +46,16 @@ class AdminSlikController extends Controller
   public function store(Request $request)
   {
     $validator = Validator::make($request->all(), [
-      'no' => 'required',
+      'no' => ['required', 'regex:/^(?:\+?62|0)[0-9]{9,13}$/'],
+      'email' => ['sometimes', 'nullable', 'email'],
       'tipe' => 'required',
       'status' => 'required',
       'ktp' => 'required|image|mimes:jpeg,png,jpg,webp|max:2048',
       'kk' => 'required|image|mimes:jpeg,png,jpg,webp|max:2048',
     ], [
       'no.required' => 'Nomor WA harus diisi.',
+      'no.regex' => 'Format Nomor WA tidak valid.',
+      'email.email' => 'Format alamat email tidak valid.',
       'tipe.required' => 'Jenis BI Checking harus dipilih.',
       'ktp.required' => 'Scan KTP harus diunggah.',
       'ktp.image' => 'File KTP harus berupa gambar.',
@@ -132,13 +135,16 @@ class AdminSlikController extends Controller
   {
     // Validasi data input
     $validator = Validator::make($request->all(), [
-      'no' => 'required',
+      'no' => ['required', 'regex:/^(?:\+?62|0)[0-9]{9,13}$/'],
+      'email' => ['sometimes', 'nullable', 'email'],
       'tipe' => 'required',
       'status' => 'required',
       'ktp' => 'image|mimes:jpeg,png,jpg,webp|max:2048',
       'kk' => 'image|mimes:jpeg,png,jpg,webp|max:2048',
     ], [
       'no.required' => 'Nomor WA harus diisi.',
+      'no.regex' => 'Format Nomor WA tidak valid.',
+      'email.email' => 'Format alamat email tidak valid.',
       'tipe.required' => 'Jenis BI Checking harus dipilih.',
       'ktp.image' => 'File KTP harus berupa gambar.',
       'ktp.mimes' => 'File KTP harus berformat jpeg, png, jpg, atau webp.',
@@ -159,7 +165,7 @@ class AdminSlikController extends Controller
       if ($request->hasFile('ktp')) {
 
         $gambarLama = public_path('assets/images/slik/ktp/' . $slik->ktp);
-        if (file_exists($gambarLama)) {
+        if (file_exists($gambarLama) && $slik->ktp) {
           unlink($gambarLama);
         }
         $gambar = $request->file('ktp');
@@ -174,7 +180,7 @@ class AdminSlikController extends Controller
       if ($request->hasFile('kk')) {
 
         $gambarLama = public_path('assets/images/slik/kk/' . $slik->kk);
-        if (file_exists($gambarLama)) {
+        if (file_exists($gambarLama) && $slik->kk) {
           unlink($gambarLama);
         }
         $gambar = $request->file('kk');
@@ -215,10 +221,10 @@ class AdminSlikController extends Controller
     try {
       $gambarKK = public_path('assets/images/slik/kk/' . $slik->kk);
       $gambarktp = public_path('assets/images/slik/ktp/' . $slik->ktp);
-      if (file_exists($gambarKK)) {
+      if (file_exists($gambarKK) && $slik->kk) {
         unlink($gambarKK);
       }
-      if (file_exists($gambarktp)) {
+      if (file_exists($gambarktp) && $slik->ktp) {
         unlink($gambarktp);
       }
       $slik->delete();
